[#5] Error for proxying unset, non-nullable association
When you try to invoke the proxy function for an association that has
not yet had a value set and is not configured to nullable, an error is
now thrown. Accessing the proxy function itself before a value is set is
legitimate so long as it is not invoked before a value is set.